pyproject.toml
[tool.poetry]
name = "public-services"
version = "1.0.0"
description = "Public services from Dan Brotsky Consulting"
authors = ["Daniel Brotsky <dev@brotsky.com>"]

[tool.poetry.dependencies]
python = "^3.8.2"
sqlalchemy = "^1.3"
alembic = "^1.3"
databases = {version = "^0.2.6", extras = ["postgresql"]}
fastapi = {version = "*", extras = ["all"]}
uvicorn = "^0.10.3"
gunicorn = "^20.0"
HALEasy = "^0.4.3"
aiohttp = "^3.6.2"
airtable-python-wrapper = "^0.13.0"
python-dateutil = "^2.8.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
coverage = "^4.5"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
